requirements.txt
pytorch
numpy
matplotlib
python-opencv
Django